backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
email-validator==1.1.3
psycopg2-binary==2.9.1
alembic==1.7.1
opencv-python==4.8.1.78
tensorflow==2.14.0
openai==1.3.5 